Re-organize example directories (#60)
* Re-organize example directories

* Move reference operation into sub-folder

* Move mask types into dedicated files

* Separate utility interface & implementation

* Resume pipeline changes in fmha_fwd.cpp

* Rename folder 'fmha_fwd' to 'fmha'

* Move more function to utils.*

* Remove 'fmha_fwd_kernel_invoker.hpp'

* Re-format files

* Move Kargs types into dedicated file

* Fix formating

* Fix compilation errors

* Avoid instantiating unused types

* Extract configurable codes

* Add missing include directive

* Instantiate template functions outside fmha_fwd.cpp

* Separate implementation files

* Merge config files

* Merge duplicated code

* Remove no-longer used file

* Unify enum name

* Extract no_mask kernel

* Further separate template specializations

* Use file(GLOB) to get file list

* Include needed config file only once

* Remove debug message

* Add comment to explain template specializations

* Move impl files under 'kernels' sub-folder

* Only include *.inc in *.inc files

* Add extra type arg to control selected kernel

* Add kernel specializations for bf16

* Switch kernel according to cmdline options

* Re-order type parameters

* Reduce loop indent level

* Instantiate launch_kernel()

* Rename source files

* Remove duplicated codes

* Remove more duplicated codes

* Clean up codes

* Rename 'FmhaMaskType' to 'FmhaMasks'

* Remove no-longer used include directive

* Move template declarations into dedicated header

* use python codegen

* modify validation logic

* format print and add smoke_test script

* modify bf16 elimit, add benchmark script

---------
