+++ example/91_tile_program/fmha/utils.hpp
@@ -1,31 +1,29 @@
 // SPDX-License-Identifier: MIT
 // Copyright (c) 2018-2023, Advanced Micro Devices, Inc. All rights reserved.
 
-#include <cassert>
-#include <functional>
+#pragma once
+
+#include <cstdint>
 #include <optional>
 #include <ostream>
-#include <random>
 #include <tuple>
 #include <utility>
 #include <vector>
 
 #include "ck/utility/span.hpp"
 
-#pragma once
-
-enum class Mode : unsigned
+enum class mode_enum
 {
-    Batch,
-    Group
+    batch = 0,
+    group
 };
 
-inline std::ostream& operator<<(std::ostream& stream, Mode mode)
+std::ostream& operator<<(std::ostream& stream, mode_enum mode)
 {
-    return stream << (mode == Mode::Batch ? "batch" : "group");
+    return stream << (mode == mode_enum::batch ? "batch" : "group");
 }
 
-inline std::vector<int32_t> to_seqstarts(ck::span<const int32_t> seqlens)
+std::vector<int32_t> to_seqstarts(ck::span<const int32_t> seqlens)
 {
     std::vector<int32_t> seqstarts = {0};
     for(int32_t seqlen : seqlens)
@@ -36,16 +34,16 @@ inline std::vector<int32_t> to_seqstarts(ck::span<const int32_t> seqlens)
     return seqstarts;
 }
 
-inline std::vector<int32_t> generate_seqlens_q(Mode mode,
-                                               unsigned count,
-                                               int32_t seqlens_q_sum,
-                                               std::optional<unsigned> seed = std::nullopt)
+std::vector<int32_t> generate_seqlens_q(mode_enum mode,
+                                        unsigned count,
+                                        int32_t seqlens_q_sum,
+                                        std::optional<unsigned> seed = std::nullopt)
 {
     assert(0 < count);
 
     std::vector<int32_t> seqlens_q(count, seqlens_q_sum);
 
-    if(mode == Mode::Group && 1 < count)
+    if(mode == mode_enum::group && 1 < count)
     {
         using size_type = std::vector<int32_t>::size_type;
 
@@ -75,26 +73,29 @@ inline std::vector<int32_t> generate_seqlens_q(Mode mode,
     return seqlens_q;
 }
 
-inline std::tuple<std::vector<int32_t>, std::vector<int32_t>> generate_seqlens_seqstarts_q(
-    Mode mode, unsigned count, int32_t seqlens_q_sum, std::optional<unsigned> seed = std::nullopt)
+std::tuple<std::vector<int32_t>, std::vector<int32_t>>
+generate_seqlens_seqstarts_q(mode_enum mode,
+                             unsigned count,
+                             int32_t seqlens_q_sum,
+                             std::optional<unsigned> seed = std::nullopt)
 {
     const std::vector<int32_t> seqlens_q = generate_seqlens_q(mode, count, seqlens_q_sum, seed);
     return std::make_tuple(seqlens_q, to_seqstarts(seqlens_q));
 }
 
-inline std::vector<int32_t> generate_seqlens_k(Mode mode,
-                                               unsigned count,
-                                               int32_t seqlens_k_sum,
-                                               ck::span<const int32_t> seqlens_q,
-                                               int32_t seqlens_q_sum,
-                                               std::optional<unsigned> seed = std::nullopt)
+std::vector<int32_t> generate_seqlens_k(mode_enum mode,
+                                        unsigned count,
+                                        int32_t seqlens_k_sum,
+                                        ck::span<const int32_t> seqlens_q,
+                                        int32_t seqlens_q_sum,
+                                        std::optional<unsigned> seed = std::nullopt)
 {
     assert(0 < count);
     assert(seqlens_q.size() == count);
 
     std::vector<int32_t> seqlens_k(count, seqlens_k_sum);
 
-    if(mode == Mode::Group && 1 < count)
+    if(mode == mode_enum::group && 1 < count)
     {
         using size_type = std::vector<int32_t>::size_type;
 
@@ -127,13 +128,22 @@ inline std::vector<int32_t> generate_seqlens_k(Mode mode,
     return seqlens_k;
 }
 
-inline std::vector<int32_t> generate_seqstarts_k(Mode mode,
-                                                 unsigned count,
-                                                 int32_t seqlens_k_sum,
-                                                 ck::span<const int32_t> seqlens_q,
-                                                 int32_t seqlens_q_sum,
-                                                 std::optional<unsigned> seed = std::nullopt)
+std::vector<int32_t> generate_seqstarts_k(mode_enum mode,
+                                          unsigned count,
+                                          int32_t seqlens_k_sum,
+                                          ck::span<const int32_t> seqlens_q,
+                                          int32_t seqlens_q_sum,
+                                          std::optional<unsigned> seed = std::nullopt)
 {
     return to_seqstarts(
         generate_seqlens_k(mode, count, seqlens_k_sum, seqlens_q, seqlens_q_sum, seed));
+}
+
+int env_get_int(const char* var_name, int default_int)
+{
+    char* v = getenv(var_name);
+    int r   = default_int;
+    if(v)
+        r = atoi(v);
+    return r;
 }
