+++ example/91_tile_program/gemm_softmax_gemm/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_example_executable(example_gemm_softmax_gemm gemm_softmax_gemm.cpp)
