+++ example/91_tile_program/gemm_softmax_gemm/gemm_softmax_gemm.cpp
@@ -13,8 +13,8 @@
 #include "ck/library/utility/host_tensor.hpp"
 #include "ck/library/utility/host_tensor_generator.hpp"
 
-#include "reference_gemm.hpp"
-#include "reference_softmax.hpp"
+#include "reference/reference_gemm.hpp"
+#include "reference/reference_softmax.hpp"
 #include "gemm_softmax_gemm.hpp"
 
 int main(int argc, char* argv[])
